--- backend/routes/AdminDeleteEventRoute.js
@@ -1,9 +1,0 @@
-import express from 'express';
-import { Router } from 'express';
-
-import { deleteEvent } from '../Controllers/AdminDeleteEventController.js';
-const router=Router();
-
-router.delete("/deleteEvent/:id",deleteEvent);
-
-export default router;
