--- backend/routes/AdminViewEventRoute.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import { Router } from 'express';
-import { getEvents } from '../Controllers/AdminViewEventController.js';
-const router=Router();
-
-router.get("/getevents",getEvents);
-
-export default router;
